Computer Player target selection improvements
added two methods to the BattleGame class that return party alignment based on the current character's turn. I.e. the player character is on the hero team, so skeletons and the like are on the monster team.

Using that, some random selections are made for friendly and enemy targets to be passed back to the Character's DoAction method along with the randomly selected move.