Implemented character health, attack damage and basic win/loss conditions
Added some extra properties to AttackAction that contain a damage value, which should vary among different attack types.

Added a MaxHealth and CurrentHealth property to Character, defining basic HP characteristics for different characters

Added logic to check if a target dies from an attack in AttackAction, and the ability to remove them from the game (removed the target's respective Party)

Added logic to check if a party has any remaining party members in BattleGame; a party loses if there are no remaining members. This condition breaks the game loop and ends the game, displaying the results to the user